Use SendToTmux in default bindings.
The Send_to_tmux function doesn't check for newlines, but SendToTmux
does. I've found that the two functions are indistinguishable in
practice for most languages, but when I'm sending to a python repl, I
really want g:tslime_ensure_trailing_newline=2 to be acknowleged.

This also fixes an oddity where sending a small sub-sexpr to a lisp
repl didn't append a newline. With this fix, it will.